fix(app): resolve all react-refresh ESLint warnings
PALETTE, Swatch, Lang and snippet() move from PalettePage.tsx to a new
PalettePage.helpers.ts (same pattern as MapPage.helpers.ts), so the
page module only exports the component and Fast Refresh works cleanly.
The react-refresh/only-export-components rule is scoped off for test
modules (*.test.{ts,tsx}, test-utils.tsx) — they never participate in
Fast Refresh, so the rule has no meaning there.
